Add option to prevent cloning if width or heigth is zero
Cloning the node can trigger additional preload requests for
descendant media elements. If a parent element is hidden, measured
width and height will be zero. Before, this always caused a fallback
to cloning the element. In some cases we might be ok though with
getting these zero values if the element is not visible anyway.